Restructure project layout and CLI entry point `dstcalc` in `pyproject.toml` for easier command execution, Updated `README.md`
