+++ lib/src/README.md
@@ -1,0 +1,11 @@
+# lib/src Directory
+
+```
+───lib/src/
+    ├── __init__.py           # Package initialization and version
+    ├── drug_database.py      # Drug database loading functions
+    ├── dst_calc.py           # Core calculation functions
+    └── supp_calc.py          # Supplementary calculation and UI functions
+```
+
+For more details, see the main `README.md` in the project root.
