+++ pyproject.toml
@@ -16,6 +16,9 @@ dependencies = [
     "tabulate>=0.9.0",
 ]
 
+[project.scripts]
+dstcalc = "app.cli.main:main"
+
 # FIXME
 # [project.scripts]
 # hello = "example:hello"
@@ -23,6 +26,10 @@ dependencies = [
 [build-system]
 requires = ["setuptools", "wheel"]
 build-backend = "setuptools.build_meta"
+
+[tool.setuptools.packages.find]
+where = ["."]
+include = ["lib*", "app*"]
 
 [tool.ruff]
 line-length = 120
