+++ setup.py
@@ -3,8 +3,7 @@ from setuptools import setup, find_packages
 setup(
     name='dstcalc',
     version='0.1.0',
-    packages=find_packages(where='src'),
-    package_dir={'': 'src'},
+    packages=find_packages(),
     install_requires=[
         'pandas',
         'shiny',
@@ -12,11 +11,6 @@ setup(
         'openpyxl',
         'pytest',
     ],
-    entry_points={
-        'console_scripts': [
-            'dstcalc=cli.main:main',
-        ],
-    },
     author='Bea Loubser',
     description='Drug Susceptibility Testing Calculator (CLI and Shiny app)',
     include_package_data=True,
