--- src/README.md
@@ -1,15 +1,0 @@
-# src Directory
-
-```
-───src/
-    ├── __init__.py           # Package initialization and version
-    ├── core/                 # Core logic and calculation functions
-    │   └── drug_database.py
-    │   └── dst_calc.py
-    ├── cli/                  # Command-line interface
-    │   └── main.py
-    ├── shinyapp/             # Shiny for Python web app
-        └── app.py    
-```
-
-For more details, see the main `README.md` in the project root.
