+++ test/app/cli/test_3.csv
@@ -1,0 +1,2 @@
+1;log1.txt;3,6;y;n;;473.4,277.23;10,5;results1.txt;6,5;10,20;final1.txt
+2;log2.txt;4 5 6;y;;;142.3,155.7,170.1;750.0;results2.txt;25.5;20;final2.txt
